feat(reth): Added botanix-cli-parsers crate
