+++ Cargo.toml
@@ -24,13 +24,14 @@ members = [
     "crates/botanix-bitcoin-checkpoint",
     "crates/botanix-btc-wallet",
     "crates/botanix-cli",
+    "crates/botanix-cli-parsers",
     "crates/botanix-comet-bft-rpc",
     "crates/botanix-data-parser",
-    "crates/botanix-storage",
-    "crates/botanix-storage-migrate",
     "crates/botanix-rpc-client",
     "crates/botanix-rpc-config",
     "crates/botanix-rpc-types",
+    "crates/botanix-storage",
+    "crates/botanix-storage-migrate",
     "crates/chain-state/",
     "crates/chainspec/",
     "crates/cli/cli/",
@@ -384,6 +385,7 @@ botanix-bitcoin-checkpoint = { path = "crates/botanix-bitcoin-checkpoint" }
 botanix-authority-metrics = { path = "crates/botanix-authority-metrics" }
 botanix-comet-bft-rpc = { path = "crates/botanix-comet-bft-rpc" }
 botanix-data-parser = { path = "crates/botanix-data-parser" }
+botanix-cli-parsers = { path = "crates/botanix-cli-parsers" }
 botanix-storage = { path = "crates/botanix-storage" }
 botanix-storage-migrate = { path = "crates/botanix-storage-migrate" }
 # frost
