+++ bin/botanix-db-cleaner/Cargo.toml
@@ -14,12 +14,12 @@ path = "src/main.rs"
 
 [dependencies]
 anyhow = "1.0.95"
+botanix-storage = { workspace = true }
 clap = { version = "4.5", features = ["derive"] }
 reth-chainspec = { workspace = true }
 reth-db = { workspace = true, features = ["mdbx"] }
 reth-node-core = { workspace = true }
 reth-provider = { workspace = true }
-botanix-storage = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 tokio = { version = "1.45.0", features = ["full"] }
