+++ bin/test-suite/Cargo.toml
@@ -30,8 +30,8 @@ botanix-authority-peg = { workspace = true, features = ["secp256k1"] }
 botanix-btc-wallet = { path = "../../crates/botanix-btc-wallet" }
 botanix-comet-bft-rpc = { workspace = true }
 botanix-data-parser = { workspace = true }
-btc-server = { path = "../btc-server" }
 botanix-storage = { workspace = true }
+btc-server = { path = "../btc-server" }
 
 # other dependencies
 bytes = "1.4.0"
