+++ crates/botanix-cli-parsers/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "botanix-cli-parsers"
+version = { workspace = true }
+edition = { workspace = true }
+rust-version = { workspace = true }
+license = { workspace = true }
+
+[dependencies]
+alloy-primitives = { workspace = true }
+eyre = { workspace = true }
+reth-cli-util = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
+thiserror = { workspace = true }
+tonic = "0.12.3"
+url = { workspace = true }
